+++ probe-rs-tools/src/bin/probe-rs/cmd/dap_server/mod.rs
@@ -9,7 +9,7 @@ mod test;
 
 use anyhow::Result;
 use probe_rs::{
-    architecture::arm::ap_v1::AccessPortError,
+    architecture::arm::ap::v1::AccessPortError,
     debug::DebugError,
     flashing::FileDownloadError,
     probe::{list::Lister, DebugProbeError},
