+++ probe-rs/src/architecture/arm/ap/memory/amba_ahb5_hprot.rs
@@ -1,5 +1,5 @@
 use crate::architecture::arm::{
-    ap_v1::{AccessPortType, ApAccess, ApRegAccess, Register},
+    ap::v1::{AccessPortType, ApAccess, ApRegAccess, Register},
     ArmError, DapAccess, FullyQualifiedApAddress, RegisterParseError,
 };
 
@@ -10,14 +10,14 @@ use super::{AddressIncrement, DataSize};
 /// The memory AP can be used to access a memory-mapped
 /// set of debug resources of the attached system.
 #[derive(Debug)]
-pub struct AmbaAhb3 {
+pub struct AmbaAhb5Hprot {
     address: FullyQualifiedApAddress,
     csw: CSW,
     cfg: super::registers::CFG,
 }
 
-impl AmbaAhb3 {
-    /// Creates a new AmbaAhb3 with `address` as base address.
+impl AmbaAhb5Hprot {
+    /// Creates a new AmbaAhb5Hprot with `address` as base address.
     pub fn new<P: DapAccess>(
         probe: &mut P,
         address: FullyQualifiedApAddress,
@@ -42,7 +42,7 @@ impl AmbaAhb3 {
     }
 }
 
-impl super::MemoryApType for AmbaAhb3 {
+impl super::MemoryApType for AmbaAhb5Hprot {
     type CSW = CSW;
 
     fn status<P: ApAccess + ?Sized>(&mut self, probe: &mut P) -> Result<CSW, ArmError> {
@@ -87,20 +87,20 @@ impl super::MemoryApType for AmbaAhb3 {
     }
 
     fn supports_only_32bit_data_size(&self) -> bool {
-        // Amba AHB3 must support word, half-word and byte size transfers.
+        // Amba AHB5 must support word, half-word and byte size transfers.
         false
     }
 }
 
-impl AccessPortType for AmbaAhb3 {
+impl AccessPortType for AmbaAhb5Hprot {
     fn ap_address(&self) -> &FullyQualifiedApAddress {
         &self.address
     }
 }
 
-impl ApRegAccess<CSW> for AmbaAhb3 {}
+impl ApRegAccess<CSW> for AmbaAhb5Hprot {}
 
-crate::attached_regs_to_mem_ap!(memory_ap_regs => AmbaAhb3);
+crate::attached_regs_to_mem_ap!(memory_ap_regs => AmbaAhb5Hprot);
 
 define_ap_register!(
     /// Control and Status Word register
@@ -108,7 +108,7 @@ define_ap_register!(
     /// The control and status word register (CSW) is used
     /// to configure memory access through the memory AP.
     name: CSW,
-    address: 0x00,
+    address_v1: 0x00,
     fields: [
         /// Is debug software access enabled.
         DbgSwEnable: bool,          // [31]
@@ -151,6 +151,8 @@ define_ap_register!(
         /// used to avoid confusion between this field and the SPIDEN signal on the authentication
         /// interface.
         SPIDEN: bool,               // [23]
+        /// `HPROT[6]`
+        HPROT6: bool,               // [15]
         /// A transfer is in progress.
         /// Can be used to poll whether an aborted transaction has completed.
         /// Read only.
@@ -175,11 +177,12 @@ define_ap_register!(
         Privileged: ((value >> 25) & 0x01) != 0,
         Data:       ((value >> 24) & 0x01) != 0,
         SPIDEN:     ((value >> 23) & 0x01) != 0,
+        HPROT6:     ((value >> 15) & 0x01) != 0,
         TrInProg:   ((value >> 7) & 0x01) != 0,
         DeviceEn:   ((value >> 6) & 0x01) != 0,
         AddrInc: AddressIncrement::from_u8(((value >> 4) & 0x03) as u8).ok_or_else(|| RegisterParseError::new("CSW", value))?,
         Size: DataSize::try_from((value & 0x07) as u8).map_err(|_| RegisterParseError::new("CSW", value))?,
-        _reserved_bits: value & 0x007F_FF08,
+        _reserved_bits: value & 0x0007_70F8,
     }),
     to: value => (u32::from(value.DbgSwEnable) << 31)
     | (u32::from(value.HNONSEC      ) << 30)
@@ -190,6 +193,7 @@ define_ap_register!(
     | (u32::from(value.Privileged   ) << 25)
     | (u32::from(value.Data         ) << 24)
     | (u32::from(value.SPIDEN       ) << 23)
+    | (u32::from(value.HPROT6       ) << 15)
     | (u32::from(value.TrInProg     ) <<  7)
     | (u32::from(value.DeviceEn     ) <<  6)
     | (u32::from(value.AddrInc as u8) <<  4)
