+++ probe-rs/src/architecture/arm/ap/v1/generic_ap.rs
@@ -1,0 +1,2 @@
+//! Generic access port
+
