+++ probe-rs/src/probe/fake_probe.rs
@@ -10,7 +10,7 @@ use probe_rs_target::ScanChainElement;
 
 use crate::{
     architecture::arm::{
-        ap_v1::memory_ap::{mock::MockMemoryAp, MemoryAp},
+        ap::v1::memory_ap::{mock::MockMemoryAp, MemoryAp},
         armv8m::Dhcsr,
         communication_interface::{
             ArmDebugState, Initialized, SwdSequence, Uninitialized, UninitializedArmProbe,
@@ -534,7 +534,7 @@ impl ArmProbeInterface for FakeArmInterface<Initialized> {
         Ok(BTreeSet::from([FullyQualifiedApAddress::v1_with_dp(dp, 1)]))
     }
 
-    fn components(&mut self, dp: DpAddress) -> Result<BTreeSet<FullyQualifiedApAddress>, ArmError> {
+    fn components(&mut self, _dp: DpAddress) -> Result<BTreeSet<FullyQualifiedApAddress>, ArmError> {
         todo!()
     }
